remove packages depending on mono being available
[Debian bug #508062]

After discussion with the Debian Mono group we decided to prune all
mono-related entries except for mono itself, which needs explicit
architecture porting.  If mono is not strictly needed for a package
to build but if it is useful only in combination with mono (e.g.
unmanaged libraries) they should introduce a build-dependency on
libmono-dev.

Rationale: In both Debian and Ubuntu mechanisms are in place to cope
with build-dependencies not being existent on some architectures.
The intended behaviour is that such packages dep-wait on the missing
packages, instead of failing on the buildds.  Like this a dependency
chain grows which will resolve itself as soon as the build-dependency
comes available on that architecture.

Signed-off-by: Philipp Kern <pkern@debian.org>
Acked-by: Luk Claes <luk@debian.org>
Acked-by: Mirco Bauer <meebey@debian.org>