Removed infrastructure to generate the web copies.
The repository moved from being something that is editied directly
to one that is updated by pushes into it, updating the web copies
after the new revisions reached it.  Thus the Makefile and the
script dumping the files from the various branches into real
files served by the web server moved into a generic binary
directory and out of this repository.