debian-edu-artwork-usplash: i386 amd64 powerpc sparc
debian-edu-artwork-usplash needs usplash and the build-dependency and
the currently single arch:any package lists those architectures
explicitly.

It needs to be excluded binary-based because there might be new packages
added to this package which are arch:any but do not have anything to do
with usplash because it's a generic artwork package.

Signed-off-by: Philipp Kern <pkern@debian.org>