[decruft] drop entries because of build-dependencies
Missing build-dependencies are handled by auto-dep-waiting mechanisms
in both Debian and Ubuntu.  Thus we can drop their entries because they
will not be tried until the build-dependencies become available.

This will increase the unbuilt percentage because the packages will be
included in the wanna-build (or equivalent) database but remain in
state dep-wait.  Furthermore some packages may still fail to build
because maintainers hardcoded an architecture list which is no longer
valid.

I still think that this is a valid optimization, though, because it
reduces the amount of p-a-s changes needed.

Signed-off-by: Philipp Kern <pkern@debian.org>